Renamed GameSettings to Player class and edited it.
